decouple script helpers
