+++ phpstan.neon
@@ -26,6 +26,7 @@ parameters:
         - tests
         - rules-tests
         - utils
+        - scripts
         - e2e/e2eTestRunnerWithCache.php
         - e2e/e2eTestRunner.php
